Remove ParticipationState INVITED
Because it conflicts with when you don't know the participation status
(see getParticipationState).

I assume the initial intended use of INVITED was to allow having a
private event and then invite users one-by-one. This should not be
mixed up with the participation-status, howeveer, as a user might be
both invited to an event and declined it. So invitation and
participation status are two separate concepts.

> Note: that there are several translations for 'Invite' which could
probably be removed now.