Cleanup: remove optional part of parameter
getParticipantCount without an explicit state could be misleading: who
exactly are we counting?

We'd be counting any user with a ParticipantStatus associated with
that event. That includes both `declined` and `accepted` statuses. A
count on the combination of this is rarely useful.

Luckily, getParticipantCount is always called with this state
parameter specified. Let's remove the optional part for clarity.